refactor: rename classes and update package structure for clarity
