--- JxInsta/src/main/java/com/errorxcode/jxinsta/AuthenticationType.java
@@ -1,18 +1,0 @@
-package com.errorxcode.jxinsta;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface AuthenticationType {
-
-    enum Method {
-        WEB_AUTH,
-        MOBILE_AUTH
-    }
-
-    Method value();
-}
