+++ src/main/java/com/errorxcode/jxinsta/auth/AuthenticationType.java
@@ -1,0 +1,18 @@
+package com.errorxcode.jxinsta.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface AuthenticationType {
+
+    enum Method {
+        WEB_AUTH,
+        MOBILE_AUTH
+    }
+
+    Method value();
+}
